Extract sales approval handlers from route modules
